Refactor custom servers
